requirements.txt
attrs==24.2.0
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
sqlparse==0.5.3
tzdata==2025.2
pytz